requirements.txt
Flask==3.0.0
pandas==2.1.3
numpy==1.26.2
scipy==1.11.4
sentence-transformers==2.2.2
faiss-cpu==1.7.4
gunicorn==21.2.0